exercises part4 done

+++ part4/bloglist/controllers/blogs.js
@@ -2,28 +2,19 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require ('../models/user')
 const jwt = require('jsonwebtoken')
-
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.startsWith('Bearer ')) {
-    return authorization.replace('Bearer ', '')
-  }
-  return null
-}
+const { userExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate('user', {username: 1, name: 1, id: 1})
       response.json(blogs)
   })
   
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', userExtractor, async (request, response) => {
   const body = request.body
-  const decodedToken = jwt.verify(request.token, process.env.SECRET)
-  if (!decodedToken.id) {
-    return response.status(401).json({error: 'token invalid'})
+  const {user} = request
+  if (!user) {
+    return response.status(401).json({ error: 'token invalid' })
   }
-
-  const fullUser = await User.findById(decodedToken.id)
 
   if (!body.title || !body.url) {
     return response.status(400).end()
@@ -34,20 +25,30 @@ blogsRouter.post('/', async (request, response) => {
     author: body.author, 
     url: body.url, 
     likes: body.likes, 
-    user: fullUser.id}
+    user: user.id}
   )
 
   const savedBlog = await blog.save()
-  fullUser.blogs = fullUser.blogs.concat(savedBlog._id)
-  await fullUser.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
 
   response.status(201).json(savedBlog)
 
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  const {user} = request
+  
+  const blog = await Blog.findById(request.params.id)
+  if ( blog.user.toString() === user.id.toString() ) {
+    await Blog.findByIdAndDelete(request.params.id)
+      response.status(204).end()
+  }
+  else{
+    response.status(401).json({ error: 'unauthorized access' })
+  }
+
+
 })
 
 blogsRouter.put('/:id', async (request, response) => {
